Day 7 solution
Took me an embarrassing amount of time to debug my part2.

I counted jokers and used that as a counter to "help" out the various
situations. However, I had the case where if you had 2 jokers, I treated
that as 2 generic cards needing "help" from jokers and counted that as
4. And the equivalent bugs in similar cases.

Took me 1:10 to solve, of which 46 minutes was part 2